requirements.txt
argcomplete @ file:///home/conda/feedstock_root/build_artifacts/argcomplete_1619128689661/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2021.10.8
cycler==0.10.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1627074840854/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
easydict==1.9
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1630267469306/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1631291095936/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1632763775723/work
ipython-genutils==0.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1610146791823/work
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1631852710654/work
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
natsort==7.1.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.1.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
ptflops==0.6.6
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==2.4.7
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-ranger==0.1.1
pytz==2021.3
pyzmq==19.0.2
scikit-learn==1.0
scipy==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
threadpoolctl==3.0.0
timm==0.4.12
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cu111
torch-optimizer==0.1.0
torchvision==0.9.0+cu111
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094704591/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
